Handle result events generically
The suite controller currently issues events for results being
available. For RHACM support, we need to do this for scansettingbindings
too, since that's the object that'll be created by RHACM as well.

this change enables such events to be created from both controllers in a
general manner.
